backend/requirements.txt
fastapi
uvicorn
playwright
google-generative-ai
python-dotenv
pydantic
